backend/requirements.txt
djangorestframework-simplejwt 
django-cors-headers 
celery 
redis 
django-celery-beat
django-otp 
pyotp